chore: rename errorBoundary.tsx to start with a capital letter
