+++ src/App.tsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import { Component } from "react";
 import SearchPage from "./pages/searchPage/SearchPage.tsx";
-import ErrorBoundary from "./components/errorBoundary/errorBoundary.tsx";
+import ErrorBoundary from "./components/errorBoundary/ErrorBoundary.tsx";
 
 class App extends Component {
   render() {
